* Fixed bug where an object would have the incorrect number of charges. + Added some extra checks inside EventableObject to prevent leakage of memory (well not leakage, but it doesn't get freed until the object is deallocated) * Fixed bug where after using a portal once you wouldn't be able to use another.

+++ src/game/GameObject.cpp
@@ -40,7 +40,7 @@ GameObject::GameObject(uint32 high, uint32 low)
 	invisibilityFlag = INVISIBILTY_FLAG_NONE;
 	spell = 0;
 	m_summoner = NULL;
-	charges = (uint32)-1;
+	charges = -1;
 	m_ritualcaster = 0;
 	m_ritualtarget = 0;
 	m_ritualmembers = NULL;
